+++ self-host/clickhouse/20260609082858_alter_bug_reports_table.sql
@@ -1,0 +1,7 @@
+-- migrate:up
+alter table bug_reports
+  comment column if exists `status` 'status of the bug report 0 (Open) or 1 (Closed)';
+
+-- migrate:down
+alter table bug_reports
+  comment column if exists `status` 'status of the bug report 0 (Closed) or 1 (Open)';